Set fugitive statusline after plugins are loaded
Creates a new function for setting plugin options and calls this on the
VimEnter event. The VimEnter event fires when enter a new Vim session,
but after all plugins are loaded.

Now we can use the `exists()` check to detect an installed plugin
instead of needing to use a bunch of `/after` directories to manage
plugin options.